added routine_log features

+++ sql/create_tables.sql
@@ -16,7 +16,7 @@ create table routine_logs(
     log_id serial not null
     ,user_id int not null default 0
     ,routine_id int not null
-    ,quantity numeric not null default 0.0
+    ,quantity double precision not null default 0.0
     ,note text not null default ''
     ,date_of_activity date not null default current_date
     ,insert_datetime timestamp with time zone not null default now()
